Remove toponyms and mapDecorations, keep Ocean Decorations #33 #30
