--- src/components/Layers/geoAssets/event.json
@@ -1,1 +1,0 @@
-{"event-20":{"x":0,"y":0,"width":1024,"height":1024}}
