+++ src/components/Layers/index.js
@@ -16,14 +16,10 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import toponymsLayer from './ToponymsLayer';
 import oceanDecorationLayer from './OceanDecorationLayer';
-import mapDecorationsLayer from './MapDecorationsLayer';
 import pinsLayer from './PinsLayer';
 
 export {
-  toponymsLayer,
   oceanDecorationLayer,
-  mapDecorationsLayer,
   pinsLayer
 };
