+++ src/models/Decor.js
@@ -16,28 +16,20 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-import { IconLayer } from '@deck.gl/layers';
-import Textures from './Textures';
+import { computed } from 'mobx';
 
-function decorationsLayer(decorations, visible, deck) {
-  return new IconLayer({
-    id: 'map-decoration-layer',
-    data: decorations,
-    visible,
-    pickable: true,
-    iconAtlas: Textures.decoration.img,
-    iconMapping: Textures.decoration.map,
-    getAngle: d => d.transform.rotate,
-    sizeScale: 3,
-    getSize: d => (deck.zoom * d.transform.scale),
-    getPosition: d => [d.geopoint[0], d.geopoint[1]],
-    getIcon: d => `decoration-${d.picId}`,
-    getColor: () => [66, 66, 66],
-    updateTriggers: {
-      getSize: deck.zoom,
-    },
-    onClick: d => console.log('decor:', d)
-  });
+import OceanDecor from './OceanDecor';
+
+export default class Decor {
+  constructor(rootStore) {
+    this.rootStore = rootStore;
+  }
+
+  @computed get oceans() {
+    return Object.values(OceanDecor).map(cur => ({
+      picId: cur.string[this.rootStore.i18n.lng],
+      geopoint: cur.geopoint,
+      size: cur.style.size,
+    }));
+  }
 }
-
-export default decorationsLayer;
